A round of cleanups + comments
